+++ artifacts/omniproject/src/pages/Settings.tsx
@@ -16,6 +16,7 @@ import { PremiumAdmin } from "../components/PremiumAdmin";
 import { LoggingSyncSettings } from "../components/settings/LoggingSyncSettings";
 import { TranslationLayer } from "../components/settings/TranslationLayer";
 import { BrokerLog } from "../components/settings/BrokerLog";
+import { A11yControls } from "../components/settings/A11yControls";
 import { DataState } from "../components/DataState";
 import { LoadingState } from "../components/LoadingState";
 import { urlFormatError } from "../lib/validation";
@@ -255,6 +256,8 @@ export function Settings() {
       </div>
 
       <PremiumAdmin />
+
+      <A11yControls />
     </div>
   );
 }